backend/requirements.txt
audiomentations==0.35.0
Faker==25.2.0
fastapi==0.111.0
guid==0.2.1
librosa==0.10.1
numpy==1.24.4
pydantic==2.7.1
pydub==0.25.1
PyJWT==2.8.0
PyJWT==2.8.0
pytz==2022.2.1
soundfile==0.12.1
SpeechRecognition==3.8.1
SQLAlchemy==2.0.30
starlette==0.37.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchvision==0.15.2+cu118
uvicorn==0.29.0
virtualenv==20.26.2
